ソート処理の統合
