+++ Client/src/store/modules/search.ts
@@ -1,6 +1,4 @@
 import { VuexModule, Module, Mutation } from "vuex-module-decorators";
-import { compare } from "natural-orderby";
-import ContentNode from "@/models/ContnetNode";
 
 export const enum SortType {
   Name,
@@ -53,25 +51,6 @@ export default class Search extends VuexModule {
       this.sorts.splice(index, 1, value);
     } else {
       this.sorts.push(value);
-    }
-  }
-
-  get compareFunc() {
-    const order = this.order;
-    switch (this.sort) {
-      case SortType.Name: {
-        const c = compare({ order: order === OrderType.Asc ? "asc" : "desc" });
-        return (x: ContentNode, y: ContentNode) => c(x.name, y.name);
-      }
-      case SortType.UpdatedAt: {
-        if (order === OrderType.Asc) {
-          return (x: ContentNode, y: ContentNode) => new Date(x.createdAt).valueOf() - new Date(y.createdAt).valueOf();
-        } else {
-          return (x: ContentNode, y: ContentNode) => new Date(y.createdAt).valueOf() - new Date(x.createdAt).valueOf();
-        }
-      }
-      default:
-        throw new RangeError("out of sorttype range");
     }
   }
 }
