+++ Client/src/views/Play.vue
@@ -55,6 +55,7 @@ import assert from "assert";
 import { assertIsDefined } from "../utilities/assert";
 import { getThumbnailPath } from "../utilities/pathUtility";
 import { delay } from "../utilities/systemUtility";
+import { compareFunc } from "../utilities/sortUtility";
 
 @Component({ components: { Player, ShakaPlayer } })
 export default class Play extends Vue {
@@ -91,7 +92,8 @@ export default class Play extends Vue {
     assertIsDefined(this.axios);
     const dir = path.dirname(this.path);
     const res = await this.axios.get<ContentNode[]>("/api/video/" + dir);
-    this.contents = res.data.filter(c => !c.isDirectory).sort(SearchModule.compareFunc);
+    const comp = compareFunc(SearchModule.sorts.find(v => v.path === this.path));
+    this.contents = res.data.filter(c => !c.isDirectory).sort(comp);
   }
 
   private isPlaying(item: ContentNode) {
